Route faults through the current telemetry instance
Move SDK fault construction into the telemetry event sink and replace process-wide session broadcasting with local router dispatch. Keep regression coverage focused on fault-specific behavior.
