+++ src/LanguageServer/Microsoft.CodeAnalysis.LanguageServer.UnitTests/Daemon/LanguageServerDaemonTests.cs
@@ -5,6 +5,7 @@
 using System.IO.Pipes;
 using System.Runtime.CompilerServices;
 using System.Threading;
+using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.LanguageServer.HostWorkspace;
@@ -112,7 +113,15 @@ public sealed class LanguageServerDaemonTests(ITestOutputHelper testOutputHelper
         var firstEvents = new RecordingEventSink();
         var secondEvents = new RecordingEventSink();
         var firstTelemetryFlushed = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
-        var firstMetrics = new RecordingMetricSink(onFlush: () => firstTelemetryFlushed.TrySetResult(true));
+        var firstFault = new InvalidOperationException("first request");
+        var reportedFirstFault = 0;
+        var firstMetrics = new RecordingMetricSink(
+            onMeasurement: () =>
+            {
+                if (Interlocked.Exchange(ref reportedFirstFault, 1) == 0)
+                    FaultReporter.ReportFault(firstFault, ErrorSeverity.General, forceDump: false);
+            },
+            onFlush: () => firstTelemetryFlushed.TrySetResult(true));
         var secondMetrics = new RecordingMetricSink();
         using var firstEventRegistration = firstTelemetry.AddEventSink(firstEvents);
         using var secondEventRegistration = secondTelemetry.AddEventSink(secondEvents);
@@ -135,6 +144,9 @@ public sealed class LanguageServerDaemonTests(ITestOutputHelper testOutputHelper
         Assert.NotNull(hover);
         Assert.True(firstMetrics.MeasurementCount > 0);
         Assert.Equal(0, secondMetrics.MeasurementCount);
+        Assert.Contains((firstFault, ErrorSeverity.General, false), firstEvents.Faults);
+        Assert.DoesNotContain(daemonEvents.Faults, fault => fault.Exception == firstFault);
+        Assert.DoesNotContain(secondEvents.Faults, fault => fault.Exception == firstFault);
 
         var daemonEventsBeforeDisconnect = daemonEvents.Events.Length;
         await first.DisposeAsync();
