+++ src/VisualStudio/Core/Def/Telemetry/Shared/TelemetryEventSink.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -14,7 +14,7 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Telemetry;
 
-internal abstract class TelemetryEventSink : IEventSink
+internal abstract partial class TelemetryEventSink : IEventSink
 {
     private sealed class Implementation : TelemetryEventSink
     {
