+++ src/Workspaces/CoreTest/Log/RoslynTelemetryTests.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.UnitTests.Logging;
 using Xunit;
@@ -23,8 +24,12 @@ public sealed class RoslynTelemetryTests
         public bool Enabled { get; set; } = true;
 
         public List<(string Kind, int BlockId)> Events { get; } = [];
+        public List<(Exception Exception, ErrorSeverity Severity, bool ForceDump)> Faults { get; } = [];
 
         public bool IsEnabled(FunctionId functionId) => Enabled && functionId == FunctionId.TestEvent_NotUsed;
+
+        public void ReportFault(Exception exception, ErrorSeverity severity, bool forceDump)
+            => Faults.Add((exception, severity, forceDump));
 
         public void Log(FunctionId functionId, LogMessage logMessage)
             => Events.Add(("Log", 0));
@@ -185,17 +190,22 @@ public sealed class RoslynTelemetryTests
     }
 
     [Fact]
-    public void NothingIsDeliveredWhenEverySinkIsDisabled()
+    public void EventsAreNotDeliveredWhenEverySinkIsDisabled()
     {
+        var telemetry = new RoslynTelemetry();
         var sink = new RecordingSink { Enabled = false };
-        using var _ = RoslynTelemetry.Current.AddEventSink(sink);
+        using var _ = telemetry.AddEventSink(sink);
 
-        RoslynTelemetry.Current.Log(FunctionId.TestEvent_NotUsed, "message");
-        using (RoslynTelemetry.Current.LogBlock(FunctionId.TestEvent_NotUsed, CancellationToken.None))
+        telemetry.Log(FunctionId.TestEvent_NotUsed, "message");
+        using (telemetry.LogBlock(FunctionId.TestEvent_NotUsed, CancellationToken.None))
         {
         }
 
         Assert.Empty(sink.Events);
+
+        var exception = new InvalidOperationException();
+        telemetry.ReportFault(exception, ErrorSeverity.Critical, forceDump: true);
+        Assert.Equal([(exception, ErrorSeverity.Critical, true)], sink.Faults);
     }
 
     [Fact]
