+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Log/IEventSink.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Threading;
+using Microsoft.CodeAnalysis.ErrorReporting;
 
 namespace Microsoft.CodeAnalysis.Internal.Log;
 
@@ -15,6 +17,11 @@ internal interface IEventSink
     /// Whether this sink will record anything for <paramref name="functionId"/>.
     /// </summary>
     bool IsEnabled(FunctionId functionId);
+
+    /// <summary>
+    /// Reports a fault. Sinks that do not support fault reporting do nothing.
+    /// </summary>
+    void ReportFault(Exception exception, ErrorSeverity severity, bool forceDump);
 
     void Log(FunctionId functionId, LogMessage logMessage);
